towerclubapp/requirements.txt
Flask==2.2.2
SQLAlchemy==1.4.27
python-dotenv==0.19.2